use float32 for Value

+++ sprint_test.go
@@ -7,7 +7,7 @@ import (
 
 func TestSprint(t *testing.T) {
 	tests := []struct {
-		arg  float64
+		arg  float32
 		want string
 	}{
 		{
